refactor: refactor comments from tech lead
