+++ src/app/main/modules/s-control/pages/form-request/screens/form-request/formatters/formatterToFormType.ts
@@ -1,4 +1,4 @@
-import { IBankTransfer, IRequestPaymentGeneral } from '../../../types/request';
+import { IBankTransfer, IRequestPaymentGeneral } from '../../../types/requestPaymentsGeneral';
 
 export function formatterToFormType(request: IRequestPaymentGeneral) {
 	const {
